refactor : extract exact-copy duplicated helpers to tests/helpers.py
Pulls the following out of individual test modules into a shared helper:
  - DB_AVAILABLE probe (was duplicated in 10 files)
  - get_db_connection() (was in 10 files as get_db or get_db_connection)
  - login_as(), logout() (was in 2 files, exact copies)
  - end_turn(page, base_url=None) (was in 3 files, near copies)
  - get_worker_id(lastname), get_controller_id(lastname)
  - load_scenario_via_admin(browser, base_url, scenario_name)

Removes ~70 lines of duplication across 11 test files. Local aliases
(get_db, _worker_id, _controller_id) preserved at import time so call
sites did not need to change.

Also fix : init_mysql.sh now bootstraps both prefixes (GAME_PREFIX and
optional GAME_PREFIX_2) at container startup. Previously only the
primary prefix had its tables created, and the parallel-games test
fixture's INSERT IGNORE on game_test2_players relied on pre-existing
tables that a fresh volume would not contain. docker-compose.yml now
sets GAME_PREFIX_2=game_test2_ so `up -d --build` initialises both.

All 180 tests pass on a fresh DB volume.

Deferred to Step 2b (after setupBase.csv): the duplicated INSERT IGNORE
gm/mechanics bootstrap blocks in 8 test modules.
