+++ tests/test_admin_csv_load_e2e.py
@@ -21,26 +21,7 @@ from conftest import (
     PHP_BASE_URL,
 )
 
-# Detect if DB is reachable — used to auto-skip DB tests when running against production
-DB_AVAILABLE = False
-try:
-    _conn = pymysql.connect(
-        host=MYSQL_HOST, port=MYSQL_PORT, user=MYSQL_USER,
-        password=MYSQL_PASSWORD, database=MYSQL_DB, connect_timeout=3,
-    )
-    _conn.close()
-    DB_AVAILABLE = True
-except Exception:
-    pass
-
-
-def get_db_connection():
-    """Get a fresh DB connection for verification."""
-    return pymysql.connect(
-        host=MYSQL_HOST, port=MYSQL_PORT, user=MYSQL_USER,
-        password=MYSQL_PASSWORD, database=MYSQL_DB,
-        charset="utf8mb4", cursorclass=pymysql.cursors.DictCursor,
-    )
+from helpers import DB_AVAILABLE, get_db_connection
 
 
 def table_row_count(table_name):
